# doc/api.yaml
openapi: 3.0.0
info:
  title: Fantastic Coffee (decaffeinated)
  version: 1.0.0
servers:
  - url: 'https://localhost:3000'
...
components:
  schemas:
    UserID:
      description: The unique id of a user, automatically generated when the account is created.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
    UserName:
      description: The username; it can be modified, but no 2 people can have the same one at the same time.
      type: string
      pattern: 'ˆ.*?$'
      minLength: 3
      maxLength: 16
    Image:
      description: The image object, made of its name and the actual image base64-encoded.
      type: string
      format: byte
      pattern: 'ˆ.*?$'
      minLength: 1
      maxLength: 10000000
    MessageID:
      description: The unique identifier of a message, automatically generated when the message is first created and sent.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
    Text:
      description: The content of a simple/plain text messsage.
      type: string
      pattern: 'ˆ.*?$'
      minLength: 1
      maxLength: 10000
    ConversationID:
      description: The unique identifier of a chat, automatically generated when the conversation is started.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
    ConversationName:
      description: The name of the conversation, which can be either the name of a group or of a person.
      type: string
      pattern: 'ˆ.*?$'
      minLength: 1
      maxLength: 20
    TimeStamp:
      description: The time at which the last message was sent.
      type: string
      format: date-time
      minLength: 20
      maxLength: 20
      readOnly: true
      example: "2017-07-21T17:32:28Z"
...
    Emoji:
      description: an emoji
      type: string
      pattern: "[\u203C-\u3299\u1F004\u1F0CF\u1F9FF]"
      minLength: 0
      maxLength: 1
      default: ""
    Reaction: 
      description: The reaction/emoji to a messsage; each person can put at most one emoji.
      type: object
      properties:
        sender:
          $ref: "#/components/schemas/UserID"
        emoji:
          $ref: "#/components/schemas/Emoji"
    User:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UserID"
        name:
          $ref: "#/components/schemas/UserName"
        image:
          $ref: "#/components/schemas/Image"
...
    Message:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/MessageID"
        conversation:
          $ref: "#/components/schemas/ConversationID"
        sender:
          $ref: "#/components/schemas/UserID"
        timestamp:
          $ref: "#/components/schemas/TimeStamp"
        checkmark:
          description: Tells you whether the message was received, read, or none of the two.
          type: string
          format: byte
          enum:
            - "sent"
            - "received"
            -  "read"
        text:
          $ref: "#/components/schemas/Text"
        image:
          $ref: "#/components/schemas/Image"
        responseTo:
          $ref: "#/components/schemas/MessageID"
        reactions:
          description: The list of all emojis (reactions) to this message.
          type: array
          minItems: 0
          maxItems: 1000
          items:
            $ref: "#/components/schemas/Reaction"
    Conversation:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ConversationID"
        name:
          $ref: "#/components/schemas/ConversationName"
        image:
          $ref: "#/components/schemas/Image"
        group:
          description: flag indicating whether the conversation is a group (true) or (false)
          type: boolean
          default: true
          readOnly: true
        lastmessage:
          $ref: "#/components/schemas/MessageID"
        participants:
          description: List of all members of a group chat. 
          type: array
          minItems: 3
          maxItems: 50
          items:
            $ref: "#/components/schemas/User"
          nullable: true
        messages:
          description: The list of all messages sent of the chat.
          type: array
          minItems: 0
          maxItems: 100000
          items:
            $ref: "#/components/schemas/Message"
  parameters:
...
    UserId:
      description: The user identifier, which must be passed for any possible operation.
      name: usrId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserID"
    ConversationId:
      description: The identifer of a conversation, used for any kind of acess to that conversation.
      name: convId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ConversationID"
    MessageId:
      description: The identifer of a message, used whenever you want to comment, respond, share, or modify that message.
      name: msgId
      in: path
      required: true
      schema: 
        $ref: "#/components/schemas/MessageID"
  securitySchemes:

    AuthenticationId:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid.
    ServerError: 
      description: Server error occurred. Please, try again in a few minutes.
    NotFoundError:
      description: The given id(s) doesn't correspond witha any user/session/message. Please try again.
    BadRequestError:
      description: You either aren't an admin to do this operation or the given input(s) are invalid.
    ItemDeleted:
      description: Item(s) succesfully deleted.
...
security:
  - AuthenticationId: []
paths:
...
  /users/{usrId}/conversations/group:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post: 
      tags: [group]
      summary: creates a new group
      description: |-
        if the operation is successful, creates a new group and add it to the list of the user's conversations.
        when creating a group, the user must specify the group's name, and at least one other member.
        if None is given, the image will be absent.
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/ConversationName"
                participants:
                  type: array
                  minItems: 3
                  maxItems: 1000
                  items:
                    $ref: "#/components/schemas/UserName"
      responses:
        "201":
          description: New group succesfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{usrId}/conversations/{convId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
    get:
      tags: [conversation]
      summary: opens the chosen conversation
      description: |-
        based on the given id, taken from the conversation's preview,
        it returns the corresponding conversation
      operationId: getConversation
      responses:
        "200":
          description: conversation successfully loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
    post: 
      tags: [group]
      summary: add new members to group
      description: |-
        adds the chosen participant to the group chat, then returns the conversation,
        with the newly updated list of participants.
      operationId: addToGroup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                participants:
                  type: array
                  minItems: 3
                  maxItems: 1000
                  items:
                    $ref: "#/components/schemas/UserName"
      responses:
        "201":
          description: New members succesfully added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  participants:
                    type: array
                    minItems: 3
                    maxItems: 1000
                    items:
                      $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [group]
      summary: leaves the group
      description: |-
        makes you leave a conversation, completely removing you from the group participants' list.
        the group will also be removed from your list of conversations and its messages deleted.
      operationId: leaveGroup
      responses:
        "204":
          $ref: "#/components/responses/ItemDeleted"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{usrId}/conversations/{convId}/messages/{msgId}: 
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
      - $ref: "#/components/parameters/MessageId"
    post:
      tags: [message]
      summary: forwards a message
      description: |-
        forwards the message to the conversation specified in the request.
        basically creating a copy of the message in the chosen conversation.
      operationId: forwardMessage
      parameters:
        - name: destId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ConversationID"
      responses:
        "201":
          description: Message forwarded succesfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [message]
      summary: deletes the message.
      description: |-
        deletes all messages that have their Ids in the list
        passed in the requestbody (up to 10 messages at once)
        if the operation succeeds, the chosen messages are cancelled
        and removed from the conversation messages' list.
        Afterwards the chat will be updated,
        showing only the remaining messages
      operationId: deleteMessage
      responses:
        "204": 
          $ref: "#/components/responses/ItemDeleted"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{usrId}/conversations/{convId}/messages/{msgId}/reactions:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
      - $ref: "#/components/parameters/MessageId"
    post: 
      tags: [message]
      summary: comments a message
      description: |-
        given a message in the chat, the user can comment it
        (a.k.a react to it) by adding the desired emoji to the message's
        reactions pool.
        If an emoji associated to your Id is already present in
        the message's reaction pool, the one in the pool will be replaced with the newly chosen one.
        If everything goes well, the updated message (with the updated reaction pool) is returned.
      operationId: commentMessage
      requestBody:
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/Emoji"
      responses:
        "201":
          description: Reaction sent or changed succesfully. 
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Reaction"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [message]
      summary: removes a reaction.
      description: |-
        checks for a reaction having the same sender id as the user,
        in which case, the reaction sent by the user for that specific message will be removed.
      operationId: uncommentMessage
      responses:
        "204": 
          $ref: "#/components/responses/ItemDeleted"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/ServerError"